Add files via upload
This is my stack implementation which is written in stack.cpp file. The basic operations of stack are used in the task 2. In task 3, a calculator is built to evaluate correct expressions and its values using stack operations and string operations. In task 3, I have used infix to postfix algorithm (as string) to evaluate the values of the correct expressions.